requirements.txt
googletrans==4.0.0rc1
opencv-python
pytesseract
torch
torchvision
pyttsx3
SpeechRecognition
google-generativeai
TTS
sounddevice
